Extract Prompt Studio textarea sizing helpers
